Move auth/user dropdown to its own component (needed for Preprints)
