+++ addon/components/navbar-auth-dropdown/style.scss
@@ -1,0 +1,5 @@
+.osf-gravatar > img {
+    border: 1px solid #CDCDCD;
+    border-radius: 13px;
+    margin-right: 5px;
+}
