+++ addon/components/navbar-auth-dropdown/template.hbs
@@ -1,0 +1,37 @@
+{{# if session.isAuthenticated }}
+    {{! TODO: Replace display name functionality if possible- for now truncate via CSS at end of label }}
+    <a href="#" class="dropdown-toggle nav-user-dropdown" data-toggle="dropdown" role="button" aria-expanded="false">
+        <span class="osf-gravatar">
+            <img src="{{gravatarUrl}}" alt="User gravatar">
+        </span> {{user.fullName}}
+        <span class="caret"></span>
+    </a>
+    <ul class="dropdown-menu" role="menu">
+        <li>
+            <a href="{{host}}profile/"><i class="fa fa-user fa-lg p-r-xs"></i> My Profile</a>
+        </li>
+        <li>
+            <a href="{{host}}support/" ><i class="fa fa-life-ring fa-lg p-r-xs"></i> Support</a>
+        </li>
+
+        <li>
+            <a href="{{host}}settings/"><i class="fa fa-cog fa-lg p-r-xs"></i> Settings</a>
+        </li>
+        <li>
+            <a onclick={{action 'logout'}}><i class="fa fa-sign-out fa-lg p-r-xs"></i> Log out</a>
+        </li>
+    </ul>
+{{else if allowLogin}}
+    {{#if institution}}  {{! TODO: How does the page know whether this is an institution view? Implement in the future }}
+        <div class="btn-group">
+            <a href="${domain}login/?campaign=institution&redirect_url=${redirect_url}" class="btn btn-info btn-top-login">
+                Sign in <span class="hidden-xs"><i class="fa fa-arrow-right"></i></span>
+            </a>
+        </div>
+    {{else}}
+        <div class="col-sm-12">
+            <a href="{{signupUrl}}" class="btn btn-success btn-top-signup m-r-xs">Sign Up</a>
+            <a {{action loginAction}} class="btn btn-info btn-top-login m-r-xs">Sign in</a>
+        </div>
+    {{/if}}
+{{/if}}
