+++ addon/components/osf-navbar/style.scss
@@ -1,9 +1,3 @@
-.osf-gravatar>img {
-    border: 1px solid #CDCDCD;
-    border-radius: 13px;
-    margin-right: 5px;
-}
-
 a.navbar-service::before {
     content: "/";
     padding-right: 8px;
