+++ addon/styles/addon.scss
@@ -1,6 +1,7 @@
 @import 'components/eosf-project-nav/style';
 @import 'components/file-browser/style';
 @import 'components/file-widget/style';
+@import 'components/navbar-auth-dropdown/style';
 @import 'components/osf-navbar/style';
 @import 'components/osf-mode-footer/style';
 @import 'components/search-dropdown/style';
