+++ app/components/navbar-auth-dropdown/component.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-osf/components/navbar-auth-dropdown/component';
